// src/db/generated/models/User.ts
export type UserModel = runtime.Types.Result.DefaultSelection<Prisma.$UserPayload>
// ...
export type UserUncheckedCreateInput = {
  id?: string
  agencyId: string
  email: string
  fullName: string
  role: $Enums.UserRole
  isActive?: boolean
  passwordHash?: string | null
  createdAt?: Date | string
  updatedAt?: Date | string
  staffMfa?: Prisma.StaffMfaUncheckedCreateNestedOneWithoutUserInput
}
// ...
export type $UserPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "User"
  objects: {
    agency: Prisma.$AgencyPayload<ExtArgs>
    staffMfa: Prisma.$StaffMfaPayload<ExtArgs> | null
  }
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    agencyId: string
    email: string
    fullName: string
    role: $Enums.UserRole
    isActive: boolean
    passwordHash: string | null
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["user"]>
  composites: {}
}